+++ imports/client/components/Tag.tsx
@@ -11,7 +11,7 @@ import OverlayTrigger from 'react-bootstrap/OverlayTrigger';
 import Popover from 'react-bootstrap/Popover';
 import { Link } from 'react-router-dom';
 import { PuzzleType } from '../../lib/schemas/puzzle';
-import { TagType } from '../../lib/schemas/tags';
+import { TagType } from '../../lib/schemas/tag';
 import { RelatedPuzzleList, sortPuzzlesByRelevanceWithinPuzzleGroup } from './RelatedPuzzleList';
 
 const PopoverPadding = {
