+++ imports/lib/models/tags.ts
@@ -1,9 +1,9 @@
 import { huntsMatchingCurrentUser } from '../../model-helpers';
-import TagsSchema, { TagType } from '../schemas/tags';
+import TagSchema, { TagType } from '../schemas/tag';
 import Base from './base';
 
 const Tags = new Base<TagType>('tags');
-Tags.attachSchema(TagsSchema);
+Tags.attachSchema(TagSchema);
 Tags.publish(huntsMatchingCurrentUser);
 
 export default Tags;
